+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## [0.3.13]() (XXXX-XX-XX)
--
+- Added side car changed to pod rotation criterium
 
 ## [0.3.12](https://github.com/arangodb/kube-arangodb/tree/0.3.12) (2019-07-04)
 - Limit source IP ranges for external services